Added function displayForecast to JS
Removed forecast HTML and created a new function called displayForecast which we are calling from the bottom of our JS. Inside the function we are selecting the element #forecast we just created and we are creating a new string called forecastHTML which has just a row so we can create a grid. We also created a new variable with 4 days in it just to test this. We are looping through each day one at a time. This is just a proof of concept. Then inside the loop we are going through each day and each time we are overriding the forecastHTML by adding what we had before and also concatenating the whole block of HTML. So every time something gets added to it. So each time we are using the HTML but we are updating the ${day}. Made sure to close the <div> for the row at the end, and we are concatenating this to the end of the existing HTML. Then we are putting this inside the forecast element we selected before.