Added 2 functions to get/display forecast from API
Updated displayTemperature function to add a call to getForecast(response.data.coord); and we will give you the coordinates. Then created function displayForecast which receives the coordinates from the displayTemperature which is the initial API call which gives us the weather of the location that we search for. Then we created an API call for the displayForecast and are giving it the Lat and Lon, the API key and the units. Then we are saying to Axios, can you go to this API Url and once you get the response go and display the forecast. Which will go into the displayForecast function and we are logging the response which is just the Array of days and each includes temp. Will work with that in the next update.