Display the forecast data using the API
Stored the response from the API inside a variable called forecast, then instead of looping through the days we had before, we are looping through the forecast variable (forecast.forEach). Which is an Array of 8 items, for each upcoming day including the current day. Then we are looking through each of them and then each of them is an object in this bunch of data. We are also getting the index from the forEach function which will include index of the current item i the Array. Then we are saying we want it to format the day of the date stamp using the function formatDay. Then we are also including the icon name inside the image URL like we did before. We are also rounding the Max/Min temperature for the forecasted days. The last thing we did is we are checking of the index is not greater than 5, and if that is the case then we are not displaying more then 6 days, it will only display the first 6. We did this so it doesn't show 8 days in a row.